talks/hexagonal-django/example/requirements.txt
attrs==16.3.0
dependencies==0.13
Django==1.10.5
